Lab: cast TerminationStatus raw to i32 for RenderTerm
ozone-app left get_raw() (u32) in Other(i32). Unblocks cef-host
compile for e2e; ozone slice can replace with a typed mapping.

+++ src-tauri/cef-host/src/app.rs
@@ -212,7 +212,7 @@ fn render_term_from_status(status: TerminationStatus) -> RenderTerm {
     } else if status == TerminationStatus::LAUNCH_FAILED {
         RenderTerm::LaunchFailed
     } else {
-        RenderTerm::Other(status.get_raw())
+        RenderTerm::Other(status.get_raw() as i32)
     }
 }
